Remove const from ImmutableDict::map_, so it can be moved out
When we pass factory lambdas to ClassWrapper::getOrCreate, it might
happen that we eventually move-construct the new Immutable{List,Dict}
object from a temporary object. In that case we want move the immer
map/vector out of the temporary object. Minor perfomance tweak.

Also add `const` to other member that can be const.